+++ components/src/Table/TableCell.js
@@ -1,35 +1,37 @@
 import React from "react";
+import PropTypes from "prop-types";
 import styled from "styled-components";
 import theme from "../theme";
 import { columnPropType, rowPropType } from "./Table.types";
 
-const getDefaultTableCellStyles = align => ({
+const StyledTableCell = styled.td(({ align }) => ({
   paddingTop: theme.space.x2,
   paddingBottom: theme.space.x2,
-  textAlign: align
-});
-
-const StyledTableCell = styled.td(({ align, isCustomCell }) => ({
+  textAlign: align,
   paddingRight: theme.space.x2,
-  "&:first-of-type": {
+  "&:first-child": {
     paddingLeft: theme.space.x2
-  },
-  ...(isCustomCell ? null : getDefaultTableCellStyles(align))
+  }
 }));
 
-const TableCell = ({ row, column }) => {
-  const isCustomCell = Boolean(column.cellRenderer);
-  const renderCellContent = (renderer, cellData) => (renderer ? renderer({ cellData, column, row }) : cellData);
-  return (
-    <StyledTableCell align={column.align} isCustomCell={isCustomCell}>
-      {renderCellContent(column.cellRenderer || column.cellFormatter, row[column.dataKey])}
-    </StyledTableCell>
+const TableCell = ({ row, column, colSpan, children }) => {
+  const cellRenderer = row.cellRenderer || column.cellRenderer;
+  const { cellFormatter } = column;
+  const isCustomCell = Boolean(cellRenderer);
+  const cellContent = cellFormatter ? cellFormatter({ cellData: children, column, row }) : children;
+  const CustomizedCell = ({ renderer, cellData, colSpan }) => (
+    <td colSpan={colSpan}>{renderer ? renderer({ cellData, column, row }) : cellData}</td>
   );
+  if (isCustomCell) {
+    return <CustomizedCell renderer={cellRenderer} cellData={children} colSpan={colSpan} />;
+  }
+  return <StyledTableCell align={column.align}>{cellContent}</StyledTableCell>;
 };
 
 TableCell.propTypes = {
   column: columnPropType.isRequired,
-  row: rowPropType.isRequired
+  row: rowPropType.isRequired,
+  colSpan: PropTypes.number.isRequired
 };
 
 export default TableCell;